Refactor code for better separation of concerns
Separate calculation logic into CalculatorLogic class.
Create ThemeLoader class to handle loading themes from YAML.
Introduce ThemeApplier class to apply themes to UI components.
Update CalculatorUI class to delegate responsibilities to CalculatorLogic and ThemeApplier.
This refactor improves code readability, maintainability, and follows the single responsibility principle.
