+++ src/main/java/com/houarizegai/calculator/ui/ThemeApplier.java
@@ -1,0 +1,71 @@
+package com.houarizegai.calculator.ui;
+
+import com.houarizegai.calculator.theme.properties.Theme;
+
+import javax.swing.*;
+
+import static com.houarizegai.calculator.util.ColorUtil.hex2Color;
+
+public class ThemeApplier {
+
+    public static void applyTheme(JFrame window, JComboBox<String> comboCalculatorType, JComboBox<String> comboTheme,
+                                  JTextField inputScreen, JButton btn0, JButton btn1, JButton btn2, JButton btn3,
+                                  JButton btn4, JButton btn5, JButton btn6, JButton btn7, JButton btn8, JButton btn9,
+                                  JButton btnPoint, JButton btnC, JButton btnBack, JButton btnMod, JButton btnDiv,
+                                  JButton btnMul, JButton btnSub, JButton btnAdd, JButton btnRoot, JButton btnLog,
+                                  JButton btnPower, JButton btnEqual, Theme theme) {
+        window.getContentPane().setBackground(hex2Color(theme.getApplicationBackground()));
+
+        comboCalculatorType.setForeground(hex2Color(theme.getTextColor()));
+        comboTheme.setForeground(hex2Color(theme.getTextColor()));
+        inputScreen.setForeground(hex2Color(theme.getTextColor()));
+        btn0.setForeground(hex2Color(theme.getTextColor()));
+        btn1.setForeground(hex2Color(theme.getTextColor()));
+        btn2.setForeground(hex2Color(theme.getTextColor()));
+        btn3.setForeground(hex2Color(theme.getTextColor()));
+        btn4.setForeground(hex2Color(theme.getTextColor()));
+        btn5.setForeground(hex2Color(theme.getTextColor()));
+        btn6.setForeground(hex2Color(theme.getTextColor()));
+        btn7.setForeground(hex2Color(theme.getTextColor()));
+        btn8.setForeground(hex2Color(theme.getTextColor()));
+        btn9.setForeground(hex2Color(theme.getTextColor()));
+        btnPoint.setForeground(hex2Color(theme.getTextColor()));
+        btnC.setForeground(hex2Color(theme.getTextColor()));
+        btnBack.setForeground(hex2Color(theme.getTextColor()));
+        btnMod.setForeground(hex2Color(theme.getTextColor()));
+        btnDiv.setForeground(hex2Color(theme.getTextColor()));
+        btnMul.setForeground(hex2Color(theme.getTextColor()));
+        btnSub.setForeground(hex2Color(theme.getTextColor()));
+        btnAdd.setForeground(hex2Color(theme.getTextColor()));
+        btnRoot.setForeground(hex2Color(theme.getTextColor()));
+        btnLog.setForeground(hex2Color(theme.getTextColor()));
+        btnPower.setForeground(hex2Color(theme.getTextColor()));
+        btnEqual.setForeground(hex2Color(theme.getBtnEqualTextColor()));
+
+        comboCalculatorType.setBackground(hex2Color(theme.getApplicationBackground()));
+        comboTheme.setBackground(hex2Color(theme.getApplicationBackground()));
+        inputScreen.setBackground(hex2Color(theme.getApplicationBackground()));
+        btn0.setBackground(hex2Color(theme.getNumbersBackground()));
+        btn1.setBackground(hex2Color(theme.getNumbersBackground()));
+        btn2.setBackground(hex2Color(theme.getNumbersBackground()));
+        btn3.setBackground(hex2Color(theme.getNumbersBackground()));
+        btn4.setBackground(hex2Color(theme.getNumbersBackground()));
+        btn5.setBackground(hex2Color(theme.getNumbersBackground()));
+        btn6.setBackground(hex2Color(theme.getNumbersBackground()));
+        btn7.setBackground(hex2Color(theme.getNumbersBackground()));
+        btn8.setBackground(hex2Color(theme.getNumbersBackground()));
+        btn9.setBackground(hex2Color(theme.getNumbersBackground()));
+        btnPoint.setBackground(hex2Color(theme.getNumbersBackground()));
+        btnC.setBackground(hex2Color(theme.getOperatorBackground()));
+        btnBack.setBackground(hex2Color(theme.getOperatorBackground()));
+        btnMod.setBackground(hex2Color(theme.getOperatorBackground()));
+        btnDiv.setBackground(hex2Color(theme.getOperatorBackground()));
+        btnMul.setBackground(hex2Color(theme.getOperatorBackground()));
+        btnSub.setBackground(hex2Color(theme.getOperatorBackground()));
+        btnAdd.setBackground(hex2Color(theme.getOperatorBackground()));
+        btnRoot.setBackground(hex2Color(theme.getOperatorBackground()));
+        btnLog.setBackground(hex2Color(theme.getOperatorBackground()));
+        btnPower.setBackground(hex2Color(theme.getOperatorBackground()));
+        btnEqual.setBackground(hex2Color(theme.getBtnEqualBackground()));
+    }
+}
